+++ rollup.config.js
@@ -36,7 +36,5 @@ export default {
 		production && terser(),
 		json()
 	],
-	experimentalCodeSplitting: true,
-	experimentalDynamicImport: true,
 	watch: { clearScreen: true }
 };