+++ src/micro/quick.svelte
@@ -1,27 +1,26 @@
 <script>
       export let link;
-
-      let linkImg;
-
-      fetch("http://localhost:4000/favicon?link=" + link.imgLink)
-            .then((r) => r.text())
-            .then((fi) => {
-                  linkImg.src = "data:image/png;base64, " + fi;
-            });
 </script>
 
 <style type="text/scss">
       .quick {
             margin: 0.25em 0.75em;
+            font-size: 0.75em;
             img {
                   width: 75px;
                   height: 75px;
-                  object-fit: cover;
+                  object-fit: contain;
                   border-radius: 15px;
+                  transition: transform 0.3s ease;
             }
             a {
                   text-decoration: none;
-                  color: #fff;
+                  color: #ddd;
+            }
+            &:hover {
+                  img {
+                        transform: scale(1.1);
+                  }
             }
       }
 </style>
@@ -30,11 +29,10 @@
       <a href={link.link}>
             <div>
                   <img
-                        bind:this={linkImg}
-                        src="./assets/link.png"
-                        onerror="this.onerror=null;this.src='{link.imgLink}';"
+                        src="https://{link.imgLink}"
+                        onerror="this.onerror=null;this.src='https://{link.imgLink}';"
                         alt={link.name} />
             </div>
-            <span style="font-size:0.75em;color:#ddd"> {link.name} </span>
+            {link.name}
       </a>
 </div>